Replace cstdints and add template is_less comparer

+++ include/common/helpers.h
@@ -1,10 +1,8 @@
 #ifndef COMMON_HELPERS_H
 #define COMMON_HELPERS_H
 
-#include <cstdint>
-
 namespace common {
-    uint64_t combine_ids(uint32_t id1, uint32_t id2);
+    unsigned long long combine_ids(unsigned long id1, unsigned long id2);
 }
 
 #endif /* COMMON_HELPERS_H */
