+++ src/helpers.cpp
@@ -1,8 +1,8 @@
 #include <common/helpers.h>
 #include <cmath>
 
-uint64_t common::combine_ids(const uint32_t id1, const uint32_t id2) {
-    uint32_t low, high;
+unsigned long long common::combine_ids(const unsigned long id1, const unsigned long id2) {
+    unsigned long low, high;
     if (id1 < id2) {
         low = id2;
         high = id1;
@@ -11,7 +11,7 @@ uint64_t common::combine_ids(const uint32_t id1, const uint32_t id2) {
         high = id2;
     }
 
-    auto dn = static_cast<uint64_t>(std::ceil(std::log10(low + 0.001)));
-    auto res = static_cast<uint64_t>(high * std::ceil(std::pow(10, dn)));
+    auto dn = static_cast<unsigned long long>(std::ceil(std::log10(low + 0.001)));
+    auto res = static_cast<unsigned long long>(high * std::ceil(std::pow(10, dn)));
     return res + low;
 }
